refactor(tauri): move GenerateTemplateOption to SelectParameter.ts
GenerateTemplateOption is generate-command-specific, so it belongs
in SelectParameter.ts alongside GenerateOptions, not in the shared
CommandOption.ts model file.
