+++ tauri/src/app/form/section/JxlsFormSection.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { SectionHelpButton } from "../../../components/dialog";
 import { ExpandButton } from "../../../components/element/ButtonIcon";
-import type { GenerateTemplateOption } from "../../../model/CommandOption";
+import type { GenerateTemplateOption } from "../../../model/SelectParameter";
 import Check from "./element/Check";
 import PlainText from "./element/PlainText";
